Fix build breaks from first impl
